refactor: better git test fixtures
