+++ __tests__/lib/fetch.test.ts
@@ -4,13 +4,17 @@ import { describe, beforeEach, afterEach, it, expect, vi, type MockInstance } fr
 import pkg from '../../package.json' with { type: 'json' };
 import { cleanAPIv1Headers, handleAPIv1Res, readmeAPIv1Fetch } from '../../src/lib/readmeAPIFetch.js';
 import { getAPIv1Mock } from '../helpers/get-api-mock.js';
-import { after, before } from '../helpers/setup-gha-env.js';
+import { githubActionsEnv } from '../helpers/git-mock.js';
 
 describe('#readmeAPIv1Fetch()', () => {
   describe('GitHub Actions environment', () => {
-    beforeEach(before);
+    beforeEach(() => {
+      githubActionsEnv.before();
+    });
 
-    afterEach(after);
+    afterEach(() => {
+      githubActionsEnv.after();
+    });
 
     it('should have correct headers for requests in GitHub Action env', async () => {
       const key = 'API_KEY';
