+++ src/lib/readmeAPIFetch.ts
@@ -7,8 +7,8 @@ import { ProxyAgent } from 'undici';
 
 import { APIv1Error } from './apiError.js';
 import config from './config.js';
-import { git } from './createGHA/index.js';
 import { getPkgVersion } from './getPkg.js';
+import { git } from './git.js';
 import isCI, { ciName, isGHA } from './isCI.js';
 import { debug, warn } from './logger.js';
